PYIC-6966: Rename pact-test-helpers lib to test-helpers
