+++ .gitignore
@@ -48,7 +48,7 @@ dev
 production
 uuid_results
 
-libs/pact-test-helpers/bin
+libs/test-helpers/bin
 local-running/bin/
 
 # User-specific secrets
