+++ lambdas/call-dcmaw-async-cri/build.gradle
@@ -25,7 +25,7 @@ dependencies {
 			libs.mockitoJunit,
 			libs.pactConsumerJunit,
 			project(path: ':libs:common-services', configuration: 'tests'),
-			project(path: ':libs:pact-test-helpers')
+			project(path: ':libs:test-helpers')
 
 	testRuntimeOnly libs.junitPlatform
 }
