+++ lambdas/user-reverification/build.gradle
@@ -21,7 +21,7 @@ dependencies {
 			libs.mockitoJunit,
 			libs.pactProviderJunit,
 			project(path: ':libs:common-services', configuration: 'tests'),
-			project(path: ':libs:pact-test-helpers')
+			project(path: ':libs:test-helpers')
 
 	testRuntimeOnly libs.junitPlatform
 }
